+++ mapmatching/__init__.py
@@ -1,2 +1,3 @@
+from .graph import GeoDigraph
 from .utils.timer import Timer, timeit
 from .matching import build_geograph, ST_Matching, STATUS