+++ mapmatching/setting.py
@@ -1,6 +1,9 @@
 """ Global config """
 from pathlib import Path
 
+IP = "192.168.202.26"
+connect= f"postgresql://postgres:pcl_A5A@{IP}:5432/gis"
+
 root = Path(__file__).parent
 DEBUG_FOLDER = root / "../debug"
 LOG_FOLDER   = root / "../log"