+++ mapmatching/utils/db.py
@@ -1,10 +1,8 @@
 import geopandas as gpd
 from sqlalchemy import create_engine
+from ..setting import connect
 
-IP = "192.168.135.16"
-connect= f"postgresql://postgres:pcl_A5A@{IP}:5432/gis"
 ENGINE   = create_engine(connect)
-
 
 def gdf_to_postgis(gdf, name, engine=ENGINE, if_exists='replace', *args, **kwargs):
     """Save the GeoDataFrame to the db